--- README.md
@@ -1,1 +1,0 @@
-# GA_Hackathon-Digital_Equity
