+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "hackathon_digital_equity",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -8,8 +8,9 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
+    "dotenv": "^16.0.3",
+    "firebase": "^9.22.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
