+++ client/src/index.css
@@ -7,4 +7,7 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
-
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
