--- package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "dotenv": "^16.0.3",
-    "firebase": "^9.22.0"
-  }
-}
